refactor(dashboard): fold DashboardCard, ToggleButton, Editable into their families
- DashboardCard -> Card.Dashboard (same markup; migrate ProjectDashboard,
  ActivityTimelineCard, SecurityOverview + test mock; delete the primitive).
- ToggleButton -> re-homed into the SegmentedControl module (same markup;
  repoint billing/time-series callers; delete the standalone dir).
- Editable -> inlined into editable-text.tsx (its only consumer); delete the
  ui/Editable primitive.
Gallery updated. Rendered output unchanged. type-check + oxlint + 1550 tests green.
