+++ client/dashboard/src/components/editable-text.tsx
@@ -1,6 +1,64 @@
 import { ReactNode, useEffect, useState } from "react";
 import { InputDialog } from "./input-dialog";
-import { Editable } from "@/components/ui/Editable";
+import { cn } from "@/lib/utils";
+import { Pencil } from "lucide-react";
+import { Text } from "@/components/ui/Text";
+
+/**
+ * Hover-to-edit overlay: blurs its children on hover and floats a "Edit" affordance
+ * (or a "Can't edit" note when disabled). Local to EditableText — its only user.
+ */
+function Editable({
+  onClick,
+  children,
+  className,
+  disabled,
+}: {
+  onClick?: () => void;
+  className?: string;
+  children: ReactNode;
+  disabled?: boolean;
+}): JSX.Element {
+  const [isHovered, setIsHovered] = useState(false);
+
+  return (
+    <div
+      className={cn("group relative cursor-pointer", className)}
+      onMouseEnter={() => setIsHovered(true)}
+      onMouseLeave={() => setIsHovered(false)}
+      onClick={() => {
+        void (!disabled && onClick?.());
+      }}
+    >
+      <div
+        className={`transition-all duration-200 ${isHovered ? "blur-xs" : ""}`}
+      >
+        {children}
+      </div>
+      {isHovered && (
+        <div className="absolute inset-0 flex items-center justify-center">
+          {disabled ? (
+            <Text muted italic>
+              Can't edit
+            </Text>
+          ) : (
+            <>
+              <Pencil className="text-muted-foreground mr-1 h-4 w-4" />
+              <Text
+                className={cn(
+                  "font-medium text-inherit",
+                  disabled && "text-muted-foreground",
+                )}
+              >
+                Edit
+              </Text>
+            </>
+          )}
+        </div>
+      )}
+    </div>
+  );
+}
 
 interface EditableTextProps {
   label: string;
