+++ client/dashboard/src/pages/security/SecurityOverview.test.tsx
@@ -53,12 +53,14 @@ vi.mock("@/components/require-scope", async () => {
   };
 });
 
-vi.mock("@/components/ui/DashboardCard", async () => {
+vi.mock("@/components/ui/Card", async () => {
   const { createElement } = await import("react");
 
   return {
-    DashboardCard: ({ children }: { children: ReactNode }) =>
-      createElement("div", null, children),
+    Card: {
+      Dashboard: ({ children }: { children: ReactNode }) =>
+        createElement("div", null, children),
+    },
   };
 });
